Move templates to correct stylesheet
